backend/requirements.txt
# Django core
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0

# Database
psycopg2-binary>=2.9,<3.0

# Environment
django-environ>=0.11,<1.0

# CORS
django-cors-headers>=4.3,<5.0

# Media (avatar)
Pillow>=10.0,<11.0

# Task Queue
celery>=5.4,<6.0
redis>=5.0,<6.0

# Docker Judge Engine
docker>=7.0.0,<8.0

# Cache
django-redis>=5.4,<6.0
